% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{c.mbmdr}
\alias{c.mbmdr}
\title{Combine MB-MDR objects}
\usage{
\method{c}{mbmdr}(..., recursive = FALSE)
}
\arguments{
\item{...}{[\code{mbmdr}]\cr
\code{mbmdr} objects to be concatenated.}

\item{recursive}{[\code{logical}]\cr
Not used.}
}
\value{
A \code{mbmdr} object of concatenated \code{mbmdr} objects.
}
\description{
Combine MB-MDR objects
}
