% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{c.mdr_models}
\alias{c.mdr_models}
\title{Combine MB-MDR objects}
\usage{
\method{c}{mdr_models}(..., recursive = FALSE)
}
\arguments{
\item{...}{[\code{mdr_models}]\cr
\code{mdr_models} objects to be concatenated.}

\item{recursive}{[\code{logical}]\cr
Not used.}
}
\value{
A \code{mdr_models} object of concatenated \code{mdr_models} objects.
}
\description{
Combine MB-MDR objects
}
\details{
The concatenated results are sorted according to the test statistics
of the models.
}
