% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.options}
\alias{check.options}
\title{Check options}
\usage{
check.options(options = getOption("mbmdr"))
}
\arguments{
\item{options}{[\code{list}]\cr
  List of MB-MDR options.}
}
\value{
Throws an error if any check fails and invisibly returns TRUE otherwise.
}
\description{
Check options of mbmdR for correct type.
}

