% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Delete old MB-MDR files}
\usage{
clean(work.dir = getwd())
}
\arguments{
\item{work.dir}{[\code{string}]\cr
The directory to look in.}
}
\value{
Returns nothing.
}
\description{
Searches for old MB-MDR analyses file in the given directory.
}
