% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topfiles.R
\name{combinePartialTopFiles}
\alias{combinePartialTopFiles}
\title{Combine partial top files}
\usage{
combinePartialTopFiles(file, trait, cpus, work.dir = getwd(),
  topfiles.prefix = file.path(work.dir, "topfiles",
  paste(basename(tools::file_path_sans_ext(file)), "top", sep = "_")),
  out = file.path(work.dir, paste(basename(tools::file_path_sans_ext(file)),
  "topfile", sep = ".")), mod = file.path(work.dir,
  paste(basename(tools::file_path_sans_ext(file)), "models", sep = ".")))
}
\arguments{
\item{file}{[\code{string}]\cr
File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
Type of trait. "binary", "continuous" or "survival".}

\item{cpus}{[\code{integer}]\cr
Sets the total amount of CPUs used in \link{createPartialTopFiles}.}

\item{work.dir}{[\code{string}]\cr
Working directory for MB-MDR. Defaults to current working directory.}

\item{topfiles.prefix}{[\code{string}]\cr
Path of partial topfiles. Defaults to <\code{work.dir}>/topfiles/<\code{file}>_top.}

\item{out}{[\code{string}]\cr
Path for saving the final topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{mod}{[\code{string}]\cr
Path for saving the models file. Defaults to <\code{work.dir}>/<\code{file}>.models.}
}
\value{
BatchJobs registry object.
}
\description{
Second step of parallel workflow of MB-MDR. Creates topfiles from multiple partial topfiles.
}

