% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{createOutput}
\alias{createOutput}
\title{Create final output file}
\usage{
createOutput(file, trait, cpus, work.dir = getwd(),
  topfile = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "topfile", sep = ".")), perm.prefix = file.path(work.dir, "permutations",
  paste(basename(file_path_sans_ext(file)), "perm", sep = "_")),
  out = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "result", sep = ".")))
}
\arguments{
\item{file}{[\code{string}]\cr
File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
Type of trait. "binary", "continuous" or "survival".}

\item{cpus}{[\code{integer}]\cr
Sets the total amount of CPUs used in \link{runPermutations}.}

\item{work.dir}{[\code{string}]\cr
Working directory for MB-MDR. Defaults to current working directory.}

\item{topfile}{[\code{string}]\cr
Path of topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{perm.prefix}{[\code{string}]\cr
Path of the permutation files. Defaults to <\code{work.dir}>/permutations/<\code{file}>_perm.}

\item{out}{[\code{string}]\cr
Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.result.}
}
\value{
BatchJobs registry object.
}
\description{
Fourth step of parallel workflow of MB-MDR. Creates final output file on one CPU.
}

