% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{resumeStep}
\alias{resumeStep}
\title{Resume parallel workflow step}
\usage{
resumeStep(file.path, file, cpus)
}
\arguments{
\item{file.path}{[\code{string}]\cr
Name for the \link{BatchJobs} \link{Registry}.}

\item{file}{[\code{string}]\cr
File path of input MB-MDR file.}

\item{cpus}{[\code{string}]\cr
How many CPUs should be used for BatchJobs?}
}
\value{
BatchJobs registry object.
}
\description{
Resumes a step of a parallel workflow of a crashed session.
}
