% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{runPermutations}
\alias{runPermutations}
\title{Create permutation files}
\usage{
runPermutations(file, trait, cpus, work.dir = getwd(),
  topfile = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "topfile", sep = ".")), out.prefix = file.path(work.dir, "permutations",
  paste(basename(file_path_sans_ext(file)), "perm", sep = "_")))
}
\arguments{
\item{file}{[\code{string}]\cr
File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
Type of trait. "binary", "continuous" or "survival".}

\item{cpus}{[\code{integer}]\cr
Sets the total amount of CPUs to be used.}

\item{work.dir}{[\code{string}]\cr
Working directory for MB-MDR. Defaults to current working directory.}

\item{topfile}{[\code{string}]\cr
Path of topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{out.prefix}{[\code{string}]\cr
Path for saving the permutation files. Defaults to <\code{work.dir}>/permutations/<\code{file}>_perm.}
}
\value{
BatchJobs registry object.
}
\description{
Third step of parallel workflow of MB-MDR. Run permutations on multiple CPUs.
}
