% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/singleThread.R
\name{singleThread}
\alias{singleThread}
\title{MB-MDR Wrapper}
\usage{
singleThread(file, trait, o, log, options)
}
\arguments{
\item{file}{[\code{string}]\cr
  File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
  Type of trait. "binary", "continuous" or "survival".}

\item{o}{[\code{string}]\cr
  Sets the output file name.}

\item{log}{[\code{string}]\cr
  Sets the log file name.}

\item{options}{[\code{list}]
  MB-MDR options set by \code{\link{configure}}.}
}
\description{
Wrapper function for external function call.
}

