% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{[.mbmdr}
\alias{[.mbmdr}
\title{Extract a subset of MDR models from a MB-MDR result object}
\usage{
\method{[}{mbmdr}(mbmdr, i)
}
\arguments{
\item{mbmdr}{[\code{mbmdr}]\cr
A \code{mbmdr} object.}

\item{i}{[\code{integer}]\cr
A vector of integers, indicating which MDR models to extract.}
}
\value{
A list of \code{mdr_model}s of class \code{mdr_models}.
}
\description{
Extract a subset of MDR models from a MB-MDR result object
}
