% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{[.mdr_models}
\alias{[.mdr_models}
\title{Extract a subset of MDR models}
\usage{
\method{[}{mdr_models}(mdr_models, i)
}
\arguments{
\item{mdr_models}{[\code{mdr_models}]\cr
A list of \code{mdr_model}s of class \code{mdr_models}.}

\item{i}{[\code{integer}]\cr
A vector of integers, indicating which MDR models to extract.}
}
\value{
A list of \code{mdr_model}s of class \code{mdr_models}.
}
\description{
Extract a subset of MDR models
}
