% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topfiles.R
\name{combinePartialTopFiles}
\alias{combinePartialTopFiles}
\title{Combine partial top files}
\usage{
combinePartialTopFiles(file, trait, cpus, reg.id = "combineTopFiles",
  work.dir = getwd(), reg.dir = file.path(work.dir, "registries", reg.id),
  topfiles.prefix = file.path(work.dir, "topfiles",
  paste0(gsub("(.+)\\\\.(.*)", "\\\\1", basename(file)), "_top")),
  out = file.path(work.dir, paste0(gsub("(.+)\\\\.(.*)", "\\\\1",
  basename(file)), ".topfile")), skip = TRUE)
}
\arguments{
\item{file}{[\code{string}]\cr
  File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
  Type of trait. "binary", "continuous" or "survival".}

\item{cpus}{[\code{integer}]\cr
  Sets the total amount of CPUs used in \link{createPartialTopFiles}.}

\item{reg.id}{[\code{string}]\cr
  Name for the \link{BatchJobs} \link{Registry}. Defaults to "combineTopFiles".}

\item{work.dir}{[\code{string}]\cr
  Working directory for MB-MDR. Defaults to current working directory.}

\item{reg.dir}{[\code{string}]\cr
  Path for saving the \link{BatchJobs} \link{Registry}. Defaults to <\code{work.dir}>/registries/<\code{reg.id}>.}

\item{topfiles.prefix}{[\code{string}]\cr
  Path of partial topfiles. Defaults to <\code{work.dir}>/topfiles/<\code{file}>_top.}

\item{out}{[\code{string}]\cr
  Path for saving the final topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{skip}{[\code{logical}]\cr
  Skip creation of a new registry if a registry is found in file.dir. Defaults to TRUE.}
}
\value{
BatchJobs registry object.
}
\description{
Second step of parallel workflow of MB-MDR. Creates topfiles from multiple partial topfiles.
}

