% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/output.R
\name{createOutput}
\alias{createOutput}
\title{Create final output file}
\usage{
createOutput(file, trait, cpus, reg.id = "output", work.dir = getwd(),
  reg.dir = file.path(work.dir, "registries", reg.id),
  topfile = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "topfile", sep = ".")), perm.prefix = file.path(work.dir, "permutations",
  paste(basename(file_path_sans_ext(file)), "perm", sep = "_")),
  out = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "result", sep = ".")), skip = TRUE)
}
\arguments{
\item{file}{[\code{string}]\cr
  File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
  Type of trait. "binary", "continuous" or "survival".}

\item{cpus}{[\code{integer}]\cr
  Sets the total amount of CPUs used in \link{runPermutations}.}

\item{reg.id}{[\code{string}]\cr
  Name for the \link{BatchJobs} \link{Registry}. Defaults to "output".}

\item{work.dir}{[\code{string}]\cr
  Working directory for MB-MDR. Defaults to current working directory.}

\item{reg.dir}{[\code{string}]\cr
  Path for saving the \link{BatchJobs} \link{Registry}. Defaults to <\code{work.dir}>/registries/<\code{reg.id}>.}

\item{topfile}{[\code{string}]\cr
  Path of topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{perm.prefix}{[\code{string}]\cr
  Path of the permutation files. Defaults to <\code{work.dir}>/permutations/<\code{file}>_perm.}

\item{out}{[\code{string}]\cr
  Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.result.}

\item{skip}{[\code{logical}]\cr
  Skip creation of a new registry if a registry is found in file.dir. Defaults to TRUE.}
}
\value{
BatchJobs registry object.
}
\description{
Fourth step of parallel workflow of MB-MDR. Creates final output file on one CPU.
}

