% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createPartialTopFiles}
\alias{createPartialTopFiles}
\title{Create partial top files}
\usage{
createPartialTopFiles(id = "partialTopFiles", work.dir = getwd(),
  reg.dir = file.path(work.dir, "registries", id), skip = TRUE)
}
\arguments{
\item{id}{[\code{string}]\cr
  Name for the \link{BatchJobs} \link{Registry}. Defaults to "partialTopFiles".}

\item{work.dir}{[\code{string}]\cr
  Working directory for MB-MDR. Defaults to current working directory.}

\item{reg.dir}{[\code{string}]\cr
  Path for saving the \link{BatchJobs} \link{Registry}. Defaults to <\code{work.dir}>/registries/<\code{id}>.}

\item{skip}{[\code{logical}]\cr
  Skip creation of a new registry if a registry is found in file.dir. Defaults to TRUE.}
}
\value{
Vector of type \code{integer} with job ids.
}
\description{
First step of parallel workflow of MB-MDR. Creates partial topfiles on multiple CPUs.
}

