% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createPartialTopFiles}
\alias{createPartialTopFiles}
\title{Maps a function over lists or vectors, adding jobs to a registry.}
\usage{
createPartialTopFiles()
}
\arguments{
\item{reg}{[\code{\link{Registry}}]\cr
  Empty Registry that will store jobs for the mapping.}
}
\value{
Vector of type \code{integer} with job ids.
}
\description{
You can then submit these jobs to the batch system.
}
\examples{
x <- 1+1
 x
}

