% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topfiles.R
\name{createPartialTopFiles}
\alias{createPartialTopFiles}
\title{Create partial top files}
\usage{
createPartialTopFiles(file, trait, cpus, reg.id = "partialTopFiles",
  work.dir = getwd(), reg.dir = file.path(work.dir, "registries", reg.id),
  out.prefix = file.path(work.dir, "topfiles",
  paste(basename(file_path_sans_ext(file)), "top", sep = "_")), skip = TRUE)
}
\arguments{
\item{file}{[\code{string}]\cr
  File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
  Type of trait. "binary", "continuous" or "survival".}

\item{cpus}{[\code{integer}]\cr
  Total amount of CPUs to be used.}

\item{reg.id}{[\code{string}]\cr
  Name for the \link{BatchJobs} \link{Registry}. Defaults to "partialTopFiles".}

\item{work.dir}{[\code{string}]\cr
  Working directory for MB-MDR. Defaults to current working directory.}

\item{reg.dir}{[\code{string}]\cr
  Path for saving the \link{BatchJobs} \link{Registry}. Defaults to <\code{work.dir}>/registries/<\code{reg.id}>.}

\item{out.prefix}{[\code{string}]\cr
  Path for saving the partial topfiles. Defaults to <\code{work.dir}>/topfiles/<\code{file}>_top.}

\item{skip}{[\code{logical}]\cr
  Skip creation of a new registry if a registry is found in file.dir. Defaults to TRUE.}
}
\value{
BatchJobs registry object.
}
\description{
First step of parallel workflow of MB-MDR. Creates partial topfiles on multiple CPUs.
}

