% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{mbmdr}
\alias{mbmdr}
\title{MB-MDR Analysis}
\usage{
mbmdr(file, trait, cpus.topfiles, cpus.permutations, work.dir = getwd(),
  prefix.topfiles = file.path(work.dir, "topfiles",
  paste(basename(file_path_sans_ext(file)), "top", sep = "_")),
  topfile = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "topfile", sep = ".")), prefix.permutations = file.path(work.dir,
  "permutations", paste(basename(file_path_sans_ext(file)), "perm", sep = "_")),
  resultfile = file.path(work.dir, paste(basename(file_path_sans_ext(file)),
  "result", sep = ".")), logfile = file.path(work.dir,
  paste(basename(file_path_sans_ext(file)), "log", sep = ".")),
  exec = "mbmdr", n.pvalues = 1000, permutations = 999,
  random.seed = NULL, group.size = 10, alpha = 0.1,
  multi.test.corr = "gammaMAXT", adjustment = "CODOMINANT", dim = "2D",
  verbose = "NONE", progressbar = "NORMAL", erase = NULL,
  erase.file = NULL, filter = NULL, filter.file = NULL,
  input.format = "MBMDR", transform = "NONE")
}
\arguments{
\item{file}{[\code{string}]\cr
  File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
  Type of trait. "binary", "continuous" or "survival".}

\item{cpus.topfiles}{[\code{integer}]\cr
  Total amount of CPUs to be used for partial topfiles.}

\item{cpus.permutations}{[\code{integer}]\cr
  Sets the total amount of CPUs to be used for permutations.}

\item{work.dir}{[\code{string}]\cr
  Working directory for MB-MDR. Defaults to current working directory.}

\item{prefix.topfiles}{[\code{string}]\cr
  Path for saving the partial topfiles. Defaults to <\code{work.dir}>/topfiles/<\code{file}>_top.}

\item{topfile}{[\code{string}]\cr
  Path for saving the final topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{prefix.permutations}{[\code{string}]\cr
  Path for saving the permutation files. Defaults to <\code{work.dir}>/permutations/<\code{file}>_perm.}

\item{resultfile}{[\code{string}]\cr
  Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.result.}

\item{logfile}{[\code{string}]\cr
  Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.log.}

\item{exec}{[\code{string}]\cr
  Default mbmdr executable. Default: "mbmdr"}

\item{n.pvalues}{[\code{integer}]\cr
  Number of p-value to compute. Default: 1000}

\item{permutations}{[\code{integer}]\cr
  Permutation amount for multiple-testing. Default: 999}

\item{random.seed}{[\code{integer} or \code{NULL}]\cr
  Random seed parameter. Default: \code{NULL} (random value)}

\item{group.size}{[\code{integer}]\cr
  Minimum group size to be statistically relevant. Default: 10}

\item{alpha}{[\code{number}]\cr
  Cutoff value for the chi-square test. Default: 0.1}

\item{multi.test.corr}{[\code{string}]\cr
  Multiple-testing correction algorithm. "NONE", "MAXT", "MINP", "RAWP",
  "STRAT1", "STRAT2" or "gammaMAXT" (default).}

\item{adjustment}{[\code{string}]\cr
  Adjust method to be used. "CODOMINANT" (default), "ADDITIVE" or "NONE".}

\item{dim}{[\code{string}]\cr
  Dimension of interactions. "1D", "2D" (defaut) or "3D".}

\item{verbose}{[\code{string}]\cr
  Verbose level. "NONE" (default), "SHORT", "MEDIUM" or "LONG".}

\item{progressbar}{[\code{string}]\cr
  Progress bar. "NONE" or "NORMAL" (default).}

\item{erase}{[\code{character} or \code{NULL}]\cr
  Vector of markers to be excluded. Default: \code{NULL}}

\item{erase.file}{[\code{string} or \code{NULL}]\cr
  File of markers to be excluded. One marker per line. Default: \code{NULL}}

\item{filter}{[\code{character} or \code{NULL}]\cr
  Analyse only the pairs composed of exactly one marker
  (for instance an environment variable) from the vector of markers names.
  Default: \code{NULL}}

\item{filter.file}{[\code{string} or \code{NULL}]\cr
  Analyse only the pairs composed of exactly one marker
  (for instance an environment variable) from the file of markers names.
  One marker per line. Default: \code{NULL}}

\item{input.format}{[\code{string}]\cr
  Input file format. "MDR" or "MBMDR" (default).}

\item{transform}{[\code{string}]\cr
  Rank transformation (continuous trait only).
  "RANK_TRANSFORM" or "NONE" (default)}
}
\value{
Throws an error if any check fails and returns TRUE otherwise.
}
\description{
Run a complete MB-MDR analysis. Necessary files and directories will be created in given working directory.
}

