% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbmdr.R
\name{mbmdr}
\alias{mbmdr}
\title{MB-MDR Analysis}
\usage{
mbmdr(formula = NULL, data = NULL, file = NULL, trait, cpus.topfiles,
  cpus.permutations, work.dir = getwd(),
  prefix.topfiles = file.path(work.dir, "topfiles",
  paste(ifelse(is.null(file), "input",
  basename(tools::file_path_sans_ext(file))), "top", sep = "_")),
  topfile = file.path(work.dir, paste(ifelse(is.null(file), "input",
  basename(tools::file_path_sans_ext(file))), "topfile", sep = ".")),
  prefix.permutations = file.path(work.dir, "permutations",
  paste(ifelse(is.null(file), "input",
  basename(tools::file_path_sans_ext(file))), "perm", sep = "_")),
  resultfile = file.path(work.dir, paste(ifelse(is.null(file), "input",
  basename(tools::file_path_sans_ext(file))), "result", sep = ".")),
  logfile = file.path(work.dir, paste(ifelse(is.null(file), "input",
  basename(tools::file_path_sans_ext(file))), "log", sep = ".")),
  modelsfile = file.path(work.dir, paste(ifelse(is.null(file), "input",
  basename(tools::file_path_sans_ext(file))), "models", sep = ".")),
  exec = "mbmdr", n.pvalues = 1000, permutations = 999,
  random.seed = as.integer(Sys.Date()), group.size = 10, num.traits = 1,
  current.trait = 1, num.covariates = 0, alpha = 0.1,
  multi.test.corr = "gammaMAXT", adjustment = "CODOMINANT",
  cov.adjustment = "RESIDUALS", dim = "2D", verbose = "MEDIUM",
  progressbar = "NORMAL", erase = NULL, erase.file = NULL,
  filter = NULL, filter.file = NULL, keep = NULL, keep.file = NULL,
  replicate.file = NULL, input.format = "MBMDR", transform = "NONE",
  bj.config = NULL, cutting_value = "-a", resume = FALSE, ...)
}
\arguments{
\item{formula}{[\code{formula}]\cr
Object of class \code{formula} describing the model to fit. See details.}

\item{data}{[\code{data.frame}]\cr
Object containing all the data.}

\item{file}{[\code{string}]\cr
File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
Type of trait. "binary", "continuous" or "survival".}

\item{cpus.topfiles}{[\code{integer}]\cr
Total amount of CPUs to be used for partial topfiles.}

\item{cpus.permutations}{[\code{integer}]\cr
Sets the total amount of CPUs to be used for permutations.}

\item{work.dir}{[\code{string}]\cr
Working directory for MB-MDR. Defaults to current working directory.}

\item{prefix.topfiles}{[\code{string}]\cr
Path for saving the partial topfiles. Defaults to <\code{work.dir}>/topfiles/<\code{file}>_top.}

\item{topfile}{[\code{string}]\cr
Path for saving the final topfile. Defaults to <\code{work.dir}>/<\code{file}>.topfile.}

\item{prefix.permutations}{[\code{string}]\cr
Path for saving the permutation files. Defaults to <\code{work.dir}>/permutations/<\code{file}>_perm.}

\item{resultfile}{[\code{string}]\cr
Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.result.}

\item{logfile}{[\code{string}]\cr
Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.log.}

\item{modelsfile}{[\code{string}]\cr
Sets the models file name. Defaults to <\code{work.dir}>/<\code{file}>.models.}

\item{exec}{[\code{string}]\cr
Default mbmdr executable. Default: "mbmdr"}

\item{n.pvalues}{[\code{integer}]\cr
Number of p-value to compute. Default: 1000}

\item{permutations}{[\code{integer}]\cr
Permutation amount for multiple-testing. Default: 999}

\item{random.seed}{[\code{integer} or \code{NULL}]\cr
Random seed parameter. Default: \code{NULL} (random value)}

\item{group.size}{[\code{integer}]\cr
Minimum group size to be statistically relevant. Default: 10}

\item{num.traits}{[\code{integer}]\cr
Amount of traits.
Default: 1}

\item{current.trait}{[\code{integer}]\cr
The selected trait.
Default: 1}

\item{num.covariates}{[\code{integer}]\cr
Amount of covariates.
Default: 0}

\item{alpha}{[\code{number}]\cr
Cutoff value for the chi-square test. Default: 0.1}

\item{multi.test.corr}{[\code{string}]\cr
Multiple-testing correction algorithm. "NONE", "MAXT", "MINP", "RAWP",
"STRAT1", "STRAT2" or "gammaMAXT" (default).}

\item{adjustment}{[\code{string}]\cr
Adjust method to be used. "CODOMINANT" (default), "ADDITIVE" or "NONE".}

\item{cov.adjustment}{[\code{string}]\cr
Covariable adjust method to be used. "RESIDUALS" (default) or "ONTHEFLY.}

\item{dim}{[\code{string}]\cr
Dimension of interactions. "1D", "2D" (defaut) or "3D".}

\item{verbose}{[\code{string}]\cr
Verbose level. "SHORT", "MEDIUM" (default) or "LONG".}

\item{progressbar}{[\code{string}]\cr
Progress bar. "NONE" or "NORMAL" (default).}

\item{erase}{[\code{character} or \code{NULL}]\cr
Vector of markers to be excluded. Default: \code{NULL}}

\item{erase.file}{[\code{string} or \code{NULL}]\cr
File of markers to be excluded. One marker per line. Default: \code{NULL}}

\item{filter}{[\code{character} or \code{NULL}]\cr
Analyse only the pairs composed of exactly one marker
(for instance an environment variable) from the vector of markers names.
Default: \code{NULL}}

\item{filter.file}{[\code{string} or \code{NULL}]\cr
Analyse only the pairs composed of exactly one marker
(for instance an environment variable) from the file of markers names.
One marker per line. Default: \code{NULL}}

\item{keep}{[\code{character} or \code{NULL}]\cr
Keep only the markers from this vector of marker names.
Default: \code{NULL}}

\item{keep.file}{[\code{string} or \code{NULL}]\cr
Keep only the markers from the given file. One marker per line.
Default: \code{NULL}}

\item{replicate.file}{[\code{string} or \code{NULL}]\cr
Used as the second stage of a discovery-replication analysis: keep only the
pairs from the given output of the first stage.
Default: \code{NULL}}

\item{input.format}{[\code{string}]\cr
Input file format. "MDR" or "MBMDR" (default).}

\item{transform}{[\code{string}]\cr
Rank transformation (continuous trait only).
"RANK_TRANSFORM" or "NONE" (default)}

\item{bj.config}{[\code{string}]\cr
Location of the configuration file to load.
Default is ".BatchJobs.conf" in the current working directory.}

\item{cutting_value}{= "-a" [\code{double"-a", "-b" or numeric}]\cr
Defines the rule, which HLO-tables should be read in at the end of mbmdr.
"-a" means: all with at least one significant value
"-b" means: all with at least one "H" and one "L" and p < .05
numeric: all with p-value below the given number ("=" excluded)}

\item{resume}{[\code{flag}]\cr
Should previous found steps of a parallel workflow resumed?}

\item{...}{[\code{any}]\cr
Additional parameter passed to and from other methods.}
}
\value{
Throws an error if any check fails and returns a \code{\link{data.table}} otherwise.
}
\description{
Run a complete MB-MDR analysis. Necessary files and directories will be created in given working directory.
}
\details{
The \code{formula} must have the following form:\cr
  \code{~trait1+trait2+...+traitnt+cov1+cov2+...+covnc+marker1+marker2+...+markernm}\cr
  or\cr
  \code{~trait1+trait2+...+cov1+cov2+.}, if there are many markers.\cr
  Here \code{trait1, trait2,..., traitnt} are the \code{nt} traits, \code{cov1, cov2, ..., covnc} are the \code{nc} covariates and \code{marker1, marker2, ..., markernm} are the \code{nm} markers, i.e. SNPs and/or environment variables to be exported to the MB-MDR file format. The first form let's you explicitly choose which variables are exported whereas the latter form exports all variables in data, but let's you re-arrange the ordering of variables. If you know that the variables in your data are correctly ordered, i.e. first the \code{nt} traits, then the \code{nc} variables and then the \code{nm} markers, you can simply provide the formula \code{~.}.\cr\cr
  WARNING: The formula interface is generally slow for big datasets. Consider to provide a correctly formatted file on the filesystem via \code{file} instead.
}

