% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleThread.R
\name{runSingleThread}
\alias{runSingleThread}
\title{MB-MDR Analysis}
\usage{
runSingleThread(file, trait, out = file.path(work.dir,
  paste0(gsub("(.+)\\\\.(.*)", "\\\\1", basename(file)), ".result")),
  reg.id = "singleThread", work.dir = getwd(),
  reg.dir = file.path(work.dir, "registries", reg.id), skip = TRUE)
}
\arguments{
\item{file}{[\code{string}]\cr
  File path of input MB-MDR file.}

\item{trait}{[\code{string}]\cr
  Type of trait. "binary", "continuous" or "survival".}

\item{out}{[\code{string}]\cr
  Sets the output file name. Defaults to <\code{work.dir}>/<\code{file}>.result.}

\item{reg.id}{[\code{string}]\cr
  Name for the \link{BatchJobs} \link{Registry}. Defaults to "singleThread".}

\item{work.dir}{[\code{string}]\cr
  Working directory for MB-MDR. Defaults to current working directory.}

\item{reg.dir}{[\code{string}]\cr
  Path for saving the \link{BatchJobs} \link{Registry}. Defaults to <\code{work.dir}>/registries/<\code{reg.id}>.}

\item{skip}{[\code{logical}]\cr
  Skip creation of a new registry if a registry is found in file.dir. Defaults to TRUE.}
}
\value{
\link{BatchJobs} \link{Registry} object.
}
\description{
Non parallel MB-MDR analysis.
}

